requirements.txt
PySimpleGUI==4.60.4
Pillow==9.5.0
pandas==2.0.2
matplotlib==3.7.1
wordcloud==1.9.2
streamlit==1.23.1